pyproject.toml
[tool.poetry]
name = "productionbot"
version = "0.4"
description = ""
authors = ["Dan <georgi.jolak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
aiodns = "3.0.0"
aiohttp = "3.9.2"
apscheduler = "3.10.1"
asyncpg = "0.27.0"
berserk-downstream = "0.11.12"
chess = "1.9.3"
colorama = "0.4.6"
coloredlogs = "15.0.1"
colormath = "3.0.0"
dateparser = "1.1.8"
discord = "2.2.3"
dill = "0.3.6"
emoji = "2.9.0"
ffmpeg = "1.4"
google-api-python-client = "2.65.0"
google-api-python-client-uritemplate = "1.4.2"
httplib2 = "0.21.0"
humanfriendly = "10.0"
icecream = "2.1.3"
langcodes = "3.3.0"
language-data = "1.1"
lichess = "0.2.8"
matplotlib = "3.6.2"
names = "0.3.0"
nest-asyncio = "1.5.6"
cairosvg = "2.7.1"
numpy = "1.23.5"
oauth2client = "4.1.3"
opencv-python = "4.8.1.78"
pillow = "10.2.0"
psutil = "5.9.5"
psycopg2-binary = "2.9.9"
pycountry = "23.12.11"
pytesseract = "0.3.10"
python-dotenv = "1.0.0"
python-telegram-bot = "20.3"
pytz = "2022.5"
pynacl = "1.5.0"
tqdm = "4.65.0"
requests = "2.31.0"
schedule = "1.2.0"
sqlalchemy = "2.0.23"
sqlalchemy-utils = "0.41.1"
surrogates = "1.0.2"
telebot = "0.0.4"
webcolors = "1.13"
websockets = "11.0.3"



[build-system]
requires = ["poetry>=1.2.2"]
build-backend = "poetry.core.masonry.api"
